requirements.txt
langchain
openai
pinecone-client
pyyaml